connect to Memo

+++ Trombone/src/classes/Memo.java
@@ -27,13 +27,13 @@ public class Memo {
 	public float getX() {
 		return x;
 	}
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 	public float getY() {
 		return y;
 	}
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
 	public int getOpacity() {
